removed vestigial openslide calls
